+++ django_webix/templates/django_webix/include/toolbar_navigation.js
@@ -1,6 +1,6 @@
 {% load django_webix_utils %}
 
-{% if form.instance.get_url_list and form.instance.get_url_list != '' %}
+{% if url_list and url_list != '' %}
     $$("{{ view.webix_view_id|default:"content_right" }}").addView({
         id: 'main_toolbar_navigation',
         view: "toolbar",
@@ -14,7 +14,7 @@
                 label: 'Torna alla lista "{{ form.instance|getattr:"_meta"|getattr:"verbose_name_plural" }}"',
                 autowidth: true,
                 click: function () {
-                    load_js("{% url form.instance.get_url_list %}");
+                    load_js("{% url url_list %}");
                 }
             }
         ]